improve folder logic

+++ internal/domain/service/folder_service.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/DevAntonioJorge/go-notes/internal/domain/models"
 	"github.com/DevAntonioJorge/go-notes/internal/domain/repository"
+	"github.com/DevAntonioJorge/go-notes/internal/domain/utils/path"
 	"github.com/DevAntonioJorge/go-notes/internal/infra/dto"
 )
 
@@ -18,7 +19,11 @@ func NewFolderService(repo *repository.Repository) *FolderService {
 	}
 }
 func (s *FolderService) SaveFolder(ctx context.Context, input dto.CreateFolderRequest) (*models.Folder, error) {
-	folder, err := models.NewFolder(input.Name, input.ParentID)
+	normalizedPath := path.NormalizePath(input.Name)
+	if !path.IsValidPath(normalizedPath) {
+		return nil, models.ErrInvalidFolderPath
+	}
+	folder, err := models.NewFolder(normalizedPath, input.ParentID)
 	if err != nil {
 		return nil, err
 	}
@@ -37,6 +42,11 @@ func (s *FolderService) UpdateFolder(ctx context.Context, input dto.UpdateFolder
 	if err != nil {
 		return nil, err
 	}
+	normalizedPath := path.NormalizePath(folder.Name)
+	if !path.IsValidPath(normalizedPath) {
+		return nil, models.ErrInvalidFolderPath
+	}
+	folder.Name = normalizedPath
 	newFolder, err := s.repo.Folder.UpdateFolder(ctx, folder)
 	if err != nil {
 		return nil, err
@@ -45,14 +55,33 @@ func (s *FolderService) UpdateFolder(ctx context.Context, input dto.UpdateFolder
 }
 
 func (s *FolderService) DeleteFolder(ctx context.Context, id string) error {
+	folder, err := s.repo.Folder.GetFolder(ctx, id)
+	if err != nil {
+		return err
+	}
+	if folder == nil {
+		return models.ErrFolderNotFound
+	}
+
+	if folder.ParentID == "" || folder.ParentID == "0" || folder.ParentID == "root" {
+		return models.ErrCannotDeleteRootFolder
+	}
+
 	return s.repo.Folder.DeleteFolder(ctx, id)
 }
 
 func (s *FolderService) GetFolders(ctx context.Context, userID string) ([]*models.Folder, error) {
+	if _, err := s.repo.User.GetUserByID(userID); err != nil {
+		return nil, models.ErrUserNotFound
+	}
 	return s.repo.Folder.GetFolders(ctx, userID)
 }
 
 func (s *FolderService) GetFolderByPath(ctx context.Context, userID, path string) (*models.Folder, error) {
+	if _, err := s.repo.User.GetUserByID(userID); err != nil {
+		return nil, models.ErrUserNotFound
+	}
+
 	return s.repo.Folder.GetFolderByPath(ctx, userID, path)
 }
